CoinbaseProExchange.jl

Available functions

Public endpoints (Coinbase account is not necessary)

CoinbaseProExchange.show_server_time — Method

show_server_time(time_type::String)

Get the API server time.

Arguments

time_type::String : "iso" (default) or "epoch" (represents decimal seconds since Unix epoch)

Example

julia> show_server_time("iso")
"2021-07-06T22:09:17.231Z"

CoinbaseProExchange.show_historical_data — Method

show_historical_data(pair::String, interval::Int64)

Fetch historic rates for a product.

Arguments

pair::String : Specify currency pair, for example "ETH-EUR"

interval::Int64 : Rates are returned in grouped buckets based on specified interval. Choose one from [60, 300 (default), 900, 3600, 21600, 86400], all in seconds.

Example

julia> show_historical_data("ETH-EUR", 3600)
300×6 DataFrame
 Row │ time                 low      high     open     close    volume  
     │ Any                  Any      Any      Any      Any      Any     
─────┼──────────────────────────────────────────────────────────────────
   1 │ 2021-06-24T11:00:00  1619.45  1650.91  1619.45  1648.84  407.623
   2 │ 2021-06-24T12:00:00  1644.02  1665.64  1648.05  1655.5   446.389
   3 │ 2021-06-24T13:00:00  1640.57  1658.99  1656.58  1651.13  295.881
   4 │ 2021-06-24T14:00:00  1639.84  1652.13  1652.13  1649.94  240.949
   5 │ 2021-06-24T15:00:00  1646.88  1692.83  1650.94  1666.79  1250.91
   6 │ 2021-06-24T16:00:00  1657.07  1694.52  1665.94  1693.9   659.553

CoinbaseProExchange.show_all_products — Method

show_all_products(currency::String)

Fetch list of all available products for a given currency.

Arguments

currency::String : "EUR", "USD" (default), "GBP" etc.

Example

julia> show_all_products("EUR")
40-element Vector{String}:
 "ETC-EUR"
 "SNX-EUR"
 "NMR-EUR"
 "SOL-EUR"
 "SKL-EUR"
 "NU-EUR"
 "XTZ-EUR"
 "UMA-EUR"

CoinbaseProExchange.show_latest_trades — Method

show_latest_trades(pair::String)

Fetch list of latest trades for a given product.

Arguments

pair::String : Specify currency pair, for example "ETH-EUR"

Example

julia> show_latest_trades("ETH-EUR")
1000×5 DataFrame
  Row │ time                      price    side    size        trade_id 
      │ Any                       Float64  String  Float64     Int64    
──────┼─────────────────────────────────────────────────────────────────
    1 │ 2021-07-06T22:23:54.963Z  1949.03  sell    0.101       22822595
    2 │ 2021-07-06T22:23:53.612Z  1948.85  sell    0.0187214   22822594
    3 │ 2021-07-06T22:23:42.616Z  1948.83  sell    0.0018276   22822593
    4 │ 2021-07-06T22:23:32.605Z  1949.17  sell    0.00285893  22822592
    5 │ 2021-07-06T22:23:32.605Z  1949.07  sell    0.006588    22822591
    6 │ 2021-07-06T22:23:29.299Z  1948.8   sell    0.0540009   22822590

CoinbaseProExchange.show_product_data — Method

show_product_data(pair::String, endpoint::String)

Fetch product data based on the selected endpoint.

Arguments

pair::String : Specify currency pair, for example "ETH-EUR"

endpoint::String : Select one from ["24hr stats" (default), "product info", "product ticker", "order book 1", "order book 2", "order book 3"]

Example

julia> show_product_data("BTC-EUR", "24hr stats")
1×6 DataFrame
 Row │ high      last      low       open      volume         volume_30day   
     │ String    String    String    String    String         String         
─────┼───────────────────────────────────────────────────────────────────────
   1 │ 29620.29  28740.89  28362.27  28642.14  1058.30490447  57346.67861744

Private endpoints (Coinbase Pro account + API keys are needed)

CoinbaseProExchange.show_all_accounts — Method

show_all_accounts(user_data::UserInfo, currencies::Vector{String})

Fetch summary of all cryptocurrency accounts associated with the given API key.

Arguments

user_data::UserInfo : API data

currencies::Vector{String} : Can be set to ["all"] or a specific set, for example ["LTC", "XTZ"]

Example

julia> show_all_accounts(user_data, ["LTC", "XTZ"])
2×7 DataFrame
 Row │ currency  balance  profile_id                         trading_enabled  id                                 hold  ⋯
     │ String    Float64  String                             Bool             String                             Float ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ LTC           0.0  4617f329-2709-453b-b95d-d14727cb…             true  eed5095d-848e-490c-8738-2f2073e7…      0 ⋯
   2 │ XTZ           0.0  4617f329-2709-453b-b95d-d14727cb…             true  21f6c731-91f7-44bf-ad9e-97cc2dfb…      0

CoinbaseProExchange.show_account_info — Method

show_account_info(user_data::UserInfo, currency::String, info_type::String)

Fetch account information for a single cryptocurrency associated with the given API key.

Arguments

user_data::UserInfo : API data

currency::String : Select one, for example ["ETH"]

info_type::String : Select one from "info", "history" or "holds"

Example

julia> show_account_info(user_data, "ETH", "history")
2×8 DataFrame
 Row │ amount               balance             created_at                   order_id                           produc ⋯
     │ String               String              String                       String                             String ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ -0.0248430100000000  0.0000000000000000  2021-05-09T00:56:06.877638Z  561bd042-9bd8-412f-a905-2a231e77…  ETH-EU ⋯
   2 │ 0.0248430100000000   0.0248430100000000  2021-05-08T23:52:37.666196Z  496f1b74-5a66-45dd-9f6e-817da994…  ETH-EU
                                                                                                       4 columns omitted

CoinbaseProExchange.show_open_orders — Method

show_open_orders(user_data::UserInfo)

Fetch list of open orders associated with the given API key.

Example

julia> show_open_orders(user_data)
1×15 DataFrame
 Row │ created_at                   executed_value      fill_fees           filled_size  id                            ⋯
     │ String                       String              String              String       String                        ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ 2021-07-04T22:15:03.793256Z  0.0000000000000000  0.0000000000000000  0.00000000   5591e17e-5f79-41f0-93d5-b455e ⋯
                                                                                                      11 columns omitted

CoinbaseProExchange.show_single_order — Method

show_single_order(order_ID::String, user_data::UserInfo)

Fetch order information for a given order ID.

Arguments

order_id::String : Order ID, can be obtained via

show_open_orders(user_data)

user_data::UserInfo : API data

Example

julia> show_single_order("14c0db51-ca17-4a8d-9d2c-aa633e703358", user_data)
1×15 DataFrame
 Row │ created_at                 executed_value      fill_fees           filled_size  id                              ⋯
     │ String                     String              String              String       String                          ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ 2021-07-04T23:11:02.0001Z  0.0000000000000000  0.0000000000000000  0.00000000   14c0db51-ca17-4a8d-9d2c-aa633e7 ⋯
                                                                                                      11 columns omitted

CoinbaseProExchange.cancel_order — Method

cancel_order(order_id::String, user_data::UserInfo)

Cancel an order with a given order ID.

Arguments

order_id::String : Order ID, can be obtained via

show_open_orders(user_data)

user_data::UserInfo : API data

Example

julia>cancel_order("5591e17e-5f79-41f0-93d5-b455ec552ecc", user_data)
"5591e17e-5f79-41f0-93d5-b455ec552ecc"

CoinbaseProExchange.cancel_all_orders — Method

cancel_all_orders(user_data::UserInfo)

Cancel all open orders from the profile associated with the API key.

Example

julia>cancel_all_orders(user_data)
1-element Vector{Any}:
 "35fa8e19-f204-40c0-b995-eb2e4e168e01"

CoinbaseProExchange.show_exchange_limits — Method

show_exchange_limits(user_data::UserInfo, currency::String)

Fetch information on the payment method transfer limit for a given currency.

Arguments

user_data::UserInfo : API data

currency::String : "ETH, "BTC", "XTZ" etc.

Example

julia> show_exchange_limits(user_data, "ETH")
5×4 DataFrame
 Row │ payment_method     max         period_in_days  remaining  
     │ String             Float64     Int64           Float64    
─────┼───────────────────────────────────────────────────────────
   1 │ ideal_deposit       14.6423                 1   14.6423
   2 │ exchange_withdraw  146.423                  1  146.423
   3 │ secure3d_buy         1.75707                7    1.75707
   4 │ credit_debit_card    0.585691               7    0.585691
   5 │ paypal_withdrawal   11.7138                 1   11.7138

CoinbaseProExchange.show_fills — Method

show_fills(user_data::UserInfo, pair::String)

Get a list of recent fills of the API key's profile for a given pair.

Arguments

user_data::UserInfo : API data

pair::String : "ETH-EUR" etc.

Example

julia> show_fills(user_data, "ETH-EUR")
7×13 DataFrame
 Row │ created_at                fee                 liquidity  order_id                           price   ⋯
     │ String                    String              String     String                             String  ⋯
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ 2021-07-04T21:54:09.902Z  0.0746268391200000  T          d275ae2b-4f34-4ce9-98a7-1147bccf…  2003.90 ⋯
   2 │ 2021-07-04T15:43:54.115Z  0.0733968750000000  T          7a019bf8-bee8-4001-bb34-e3d6f3e6…  1957.25
   3 │ 2021-07-04T15:42:35.808Z  0.0597014279730000  T          2a8814f1-76d3-4fa2-bb0a-6f25e050…  1956.42
   4 │ 2021-07-04T15:28:06.005Z  0.0599999776075000  T          9971744a-d308-4481-950d-e2ada197…  1953.35